requirements.txt
absl-py==0.2.1
aniso8601==3.0.0
astor==0.6.2
astroid==1.6.4
beautifulsoup4==4.6.0
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Cors==3.0.4
Flask-RESTful==0.3.6
gast==0.2.0
grpcio==1.12.0
gunicorn==19.8.1
h5py==2.7.1
html5lib==0.9999999
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.6
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.3
opencv-python==3.4.0.14
Pillow==5.1.0
protobuf==3.5.2.post1
pylint==1.9.1
pytz==2018.4
PyYAML==3.12
requests==2.18.4
scipy==1.1.0
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
tqdm==4.23.3
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
yapf==0.22.0
